Initial commit
I have written v1.0.0 of my base converter. It can convert numbers
between 0 and 999,999,999 from bases between 2 and 36. It can also
output the result into either stdout or a file, if specified. It may
have bugs / memory leaks that I haven't found, but for now the program
runs correctly and that is good enough. In the future, I would like to
increase the range of numbers that the program accepts as well as
support for negative numbers.